Adding Relative path to routes
you need to add / while mentioning the routes else it assumes to the exisiting path it will add the routes. It perfectly works fine if don add / before the route and add importactivated routes and add relative path. Activated route tells angular the exact location of where we are in the project