Pssing the router parameters
add the routing parameters in the app.module.ts , add the routeLink in the anchor tag and then add the routing path. In the server.ts file get the snapshot of the parameter passed . To get the updated value of the parameter add subscribe method. Note: If whole url is a string and if u want to get the value of a parameter which is a number then just add + in front the params and declaration of id