routing practice
passing a query parameter as a ternary operator in the url. In the server.component add a click function and call the function and add router navigate with the route path and add relativeTo (will tell angular the exact location where we are). On click of the button we are navigated to the EditServer component we are subscribing to the query parameter