Restricting the routes by CanActive
Can activate checks for the authentication in the background before the component is been rendered. Create auth.service and authguard service one is service other is the service injectable. Add the main service to the app.module.ts providers. the main service to contain the login, logout and authentication with promise. The injected service has an observable that stores the boolean value and blocks the route if not authenticated